Move signal processing from utils.data_utils to processing
